Extract global state store helpers
